[PrepTask] Modify DT PE image to load capture script on startup
The DT PE image capture preparation task modifies DISMTools PE images to load the image capture script on startup, in a fashion similar to HotInstall's, while adapting it to our "Tasker" infrastructure. To work nicely with Tasker, some changes were made to the PT Base and the interfaces it implements, mainly to add support for recursive folder copies. Everything else is implemented at the scope of this PT.

This has worked (first try!) under specific conditions as part of the test environment. To try this (IN ITS CURRENT STATE AS OF THIS COMMIT) using a disc image created by DT:

1. Copy the DT PE boot image (boot.wim) to the "sources" folder in the root of your drive* (create that folder if it doesn't exist). Optionally, remove read-only attributes, if present (the PT will do this automatically anyway)
2. Copy the "Boot" folder of the ISO to the root of your drive*
3. Invoke the sysprep prep tool with, at least, flags `/test /dt_capture`. DO NOT SCREW THINGS UP WHEN TYPING THE FLAGS!!!

* the root of your drive means the root of the drive the Sysprep prep tool has been cloned to

Why you shouldn't mess with the flags...
- The PT will only run if you pass, at least, `/dt_capture`. You can pass other arguments alongside that one
- PASS `/test` AS AN ARGUMENT. Otherwise, the PT will modify your system's Boot Configuration Data (BCD) and will make your computer boot to a broken "Image Capture" boot entry. If that happens, get ready to recover your BCD (it should be easy because you can still boot to your perfectly-working Windows install within the 3-second timeframe). Follow the instructions at the bottom of the commit description to remove any broken "Image Capture" boot entries.

No progress output for this task is displayed yet. That will be fixed in a future change. Other things that need to be ironed out are:

- Documentation still needs to be added
- BCD modifications need to be completed
- Testing with actual systems (in audit mode) needs to be done
- DynaLog needs to be implemented

All the required scripts in DISMTools for this to work have already been prepared in 0.7.2 Preview 2.

POSSIBLE RECOVERY OPTIONS FOR BCD:

To restore the BCD so it always boots to your Windows install instead of the broken "Image Capture" entry, follow these steps...

1. Load CMD as admin
2. Get BCD entries (bcdedit) and grab the GUID of the image capture environment
3. Run `bcdedit /delete <GUID> /f` (where <GUID> is the GUID you grabbed). You may be able to skip step 2 by running `bcdedit /delete {default} /f` (this may work, but I'm not sure)
4. Run `bcdedit /default {current}` to make the current boot entry the default one in the BCD

!!!!!!!!!! DO NOT FOLLOW THESE INSTRUCTIONS IF THE BCD HAS NOT BEEN MODIFIED. DO NOT. OTHERWISE, YOUR SYSTEM WILL STOP BOOTING TO YOUR WINDOWS INSTALL !!!!!!!!!!

CLEANING UP...

It is recommended that you delete the `$DISMTOOLS.~BT` and `$DISMTOOLS.~WS` folders manually in every test run. Neither the PT nor the updated scripts in DISMTools will remove these. In HotInstall, it does not really matter as the partitions are deleted and reformatted, but, in here, this may well cause the PT to fail in subsequent runs. This is also expected to be fixed in the future (at least in the PT).